Only use FileListingBlock, remove all/rename necessary parts of ListingBlock.
