+++ ftw/simplelayout/lawgiver.zcml
@@ -13,7 +13,6 @@
     <lawgiver:map_permissions
         action_group="edit"
         permissions="ftw.simplelayout: Add File,
-                     ftw.simplelayout: Add ListingBlock,
                      ftw.simplelayout: Add FileListingBlock,
                      ftw.simplelayout: Add TextBlock,
                      ftw.simplelayout: Add MapBlock,
