--- ftw/simplelayout/profiles/default/types/ftw.simplelayout.ListingBlock.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0"?>
-<object name="ftw.simplelayout.ListingBlock"
-        meta_type="Dexterity FTI"
-        xmlns:i18n="http://xml.zope.org/namespaces/i18n"
-        i18n:domain="ftw.simplelayout" >
-
-    <!-- Basic metadata -->
-    <property name="title" i18n:translate="">ListingBlock</property>
-    <property name="description" i18n:translate="">Use this block for File or listings or galleries</property>
-    <property name="icon_expr"></property>
-    <property name="allow_discussion">False</property>
-    <property name="global_allow">False</property>
-    <property name="filter_content_types">True</property>
-    <property name="allowed_content_types">
-        <element value="ftw.simplelayout.File" />
-    </property>
-
-    <!-- schema interface -->
-    <property name="schema">ftw.simplelayout.contents.listingblock.IListingBlockSchema</property>
-
-    <!-- class used for content items -->
-    <property name="klass">ftw.simplelayout.contents.listingblock.ListingBlock</property>
-
-    <!-- add permission -->
-    <property name="add_permission">ftw.simplelayout.AddListingBlock</property>
-
-    <!-- enabled behaviors -->
-    <property name="behaviors">
-        <element value="plone.app.content.interfaces.INameFromTitle" />
-        <element value="ftw.simplelayout.interfaces.ISimplelayoutBlock" />
-    </property>
-
-    <!-- View information -->
-    <property name="default_view">block_view</property>
-    <property name="default_view_fallback">False</property>
-    <property name="view_methods">
-        <element value="block_view"/>
-    </property>
-
-    <!-- Method aliases -->
-    <alias from="(Default)" to="(dynamic view)"/>
-    <alias from="edit" to="@@edit"/>
-    <alias from="sharing" to="@@sharing"/>
-    <alias from="view" to="(selected layout)"/>
-
-    <!-- Actions -->
-    <action
-        action_id="view"
-        title="View"
-        category="object"
-        condition_expr=""
-        url_expr="string:${object_url}"
-        visible="True">
-        <permission value="View"/>
-    </action>
-
-    <action
-        action_id="edit"
-        title="Edit"
-        category="object"
-        condition_expr=""
-        url_expr="string:${object_url}/edit"
-        visible="True">
-        <permission value="Modify portal content"/>
-    </action>
-
-</object>
